APIAutomation/requirements.txt
requests==2.31.0
pytest==7.4.3
allure-pytest==2.13.2
Faker==19.13.0
pytest-xdist 